Updated Henry with details of user and method in alert
Also fixed a bug caused by ignoring acl on create bucket events -
I thought these were always passing an ACL of blank and that they would
be followed by a call to PutBucketAcl, and so a CreateBucket could
itself be ignored, but this is only the case when creating from the AWS
Management Console. When using CLI or SDK the CreateBucket permission is
applied in a single command, so Cloudwatch Event parsing needs to extend
to CreateBucket as well as PutBucketAcl.